+++ src/carriers/mpi_carrier/src/MpiP2PStream.cpp
@@ -17,7 +17,7 @@ using namespace yarp::os::impl;
 /////////////////////////////////////////////////
 // InputStream
 
-int MpiP2PStream::read(const Bytes& b) {
+ssize_t MpiP2PStream::read(const Bytes& b) {
     if (readAvail == 0) {
         // get new data
         reset();
