+++ src/carriers/xmlrpc_carrier/XmlRpcStream.h
@@ -93,7 +93,7 @@ public:
 
     virtual void write(const Bytes& b);
 
-    virtual int read(const Bytes& b);
+    virtual ssize_t read(const Bytes& b);
     //{
     //  return delegate->getInputStream().read(b);
     //}
